backend/requirements.txt
flask
flask-cors
google-genai
pymysql
python-dotenv
requests